Move `logger_factory.resque_worker` to CommandBusServiceProvider
